1.7.2.0 initial update

+++ ExtractorSharp/Command/FileCommand/SplitFile.cs
@@ -1,46 +1,47 @@
-﻿using ExtractorSharp.Core;
-using ExtractorSharp.Core.Lib;
-using ExtractorSharp.Data;
-using System.Collections.Generic;
-using System.IO;
-using System.Text.RegularExpressions;
+﻿using System.Collections.Generic;
+using ExtractorSharp.Core.Coder;
+using ExtractorSharp.Core.Command;
+using ExtractorSharp.Core.Composition;
+using ExtractorSharp.Core.Model;
 
-namespace ExtractorSharp.Command.ImgCommand {
-    class SplitFile : ICommand {
+namespace ExtractorSharp.Command.FileCommand {
+    internal class SplitFile : ICommand, IFileFlushable {
+        private Album[] _array;
 
-        private Album[] Array;
+        private List<Album> _list;
 
-        private List<Album> List;
-
-        private IConnector Connector => Program.Connector;
-
+        private static IConnector Connector => Program.Connector;
 
         public bool CanUndo => true;
 
-        public bool IsChanged => false;
-
-        public bool IsFlush => true;
+        public bool IsChanged => false;       
 
         public string Name => "SplitFile";
 
         public void Do(params object[] args) {
-            Array = args as Album[];
-            List = new List<Album>();
-            foreach (var al in Array) {
-                var arr = Npks.SplitFile(al);
+            _array = args as Album[];
+            _list = new List<Album>();
+            if (_array == null) {
+                return;
+            }
+            foreach (var al in _array) {
+                var arr = NpkCoder.SplitFile(al);
                 Connector.RemoveFile(al);
                 Connector.AddFile(false, arr);
-                List.AddRange(arr);
+                _list.AddRange(arr);
             }
         }
 
         public void Redo() {
-            Do(Array);
+            Do(_array);
         }
 
         public void Undo() {
-            Connector.RemoveFile(List.ToArray());
-            Connector.AddFile(false, Array);
+            Connector.RemoveFile(_list.ToArray());
+            if (_array == null) {
+                return;
+            }
+            Connector.AddFile(false, _array);
         }
     }
 }
